requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.2
django-appconf==1.0.5
django-bootstrap-icons==0.8.3
django-bootstrap5==23.3
django-filter==23.2
django-image-cropping==1.7
django-initials-avatar==0.0.6
django-json-widget==1.1.1
django-modeltranslation==0.18.10
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
easy-thumbnails==2.8.5
exceptiongroup==1.1.1
flower==2.0.0
future==0.18.3
h11==0.14.0
humanize==4.6.0
idna==3.4
kombu==5.3.1
Markdown==3.4.3
outcome==1.2.0
phonenumbers==8.13.14
Pillow==9.5.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sorl-thumbnail==12.9.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tornado==6.3.2
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.0
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
